// src/app/models/api-request-models/medco-node/api-sequence-of-events/api-i2b2-which-date.ts
/**
 * Copyright 2021 CHUV
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
export enum ApiI2b2WhichDate {
  startdate = 'STARTDATE',
  enddate = 'ENDDATE'
}